requirements-dev.txt
# Lambda function dependencies
boto3>=1.34
requests>=2.31
openai>=1.0
Pillow>=10.0

# Testing
pytest>=7.0
pytest-mock>=3.0
moto[dynamodb,s3,secretsmanager,sqs]>=5.0

# Code quality
ruff>=0.8.0
bandit>=1.7
pre-commit>=3.0
